### 2023-10-01 #5911 @editorialbot
[PRE REVIEW]: Shapelets: A Python package implementing shapelet functions and their applications
**Submitting author:** @nasserma (<a href="http://orcid.org/0000-0002-1772-0376">Nasser Mohieddin Abukhdeir</a>)
**Repository:** https://github.com/uw-comphys/shapelets
**Branch with paper.md** (empty if default branch): publications
**Version:** v0.1
