### 2023-10-02 36aa283
Significant changes to installation instructions in README.md (#9)
