+++ README.md
@@ -27,43 +27,55 @@ In particular, the `shapelets` package provides documentation and code for four 
 Applications implemented in the package include those from astronomy (galactic image decomposition and reconstruction) [[A. Refregier (2003)](https://doi.org/10.1046/j.1365-8711.2003.05901.x), [R. Massey (2005)](https://doi.org/10.48550/arXiv.astro-ph/0408445)] and shapelet-based methods to quantify order from self-assembly nanostructured surface imaging [[R. Suderman (2015)](http://dx.doi.org/10.1103/PhysRevE.91.033307), [T. Akdeniz (2018)](https://doi.org/10.1088/1361-6528/aaf353), [M. P. Tino (in revision)](https://github.com/uw-comphys/shapelets)]. 
 
 
-## Getting Started
+## Installation Instructions
 
-### Installation Instructions
+Installing `shapelets` highly depends on the OS of your machine. Please follow the appropriate guide depending on your OS (Linux, Mac OS, Windows). 
 
-To install `shapelets`, you should clone the repository onto your machine in the desired directory. This can be done using the Github GUI, but for simplicity it is preferred to use your terminal / command line interface, outlined below.
+### Linux
 
-1. In your desired directory, type `git clone https://github.com/uw-comphys/shapelets.git`
-2. Ensure that `pip3` is installed,
-    * Linux - `sudo apt install python3-pip`
-    * macOS - install [Homebrew](https://brew.sh/) via terminal. Then `brew install python3`, which also installs `pip3`
-    * Windows - we recommend using WSL (Windows Subsystem for Linux), please follow [**here**](#installing-shapelets-on-wsl-windows-subsystem-for-linux) before proceeding
-3. To install the `shapelets` package, type `pip3 install .` from the topmost package directory (i.e., /home/.../shapelets) 
-4. To ensure correct installation, type `shapelets-test`, which should trigger the [custom command](#custom-commands) to initiate unit testing.
+1. Update your Ubuntu system via `sudo apt-get update`
+2. Install `git`, `python3`, and `pip` via `sudo apt-get install git python3 python3-pip`
+3. Clone the `shapelets` repository via `git clone https://github.com/uw-comphys/shapelets.git`
+4. To install `shapelets`, navigate to your /shapelets directory, then type `pip3 install .`
+5. To ensure correct installation, type `shapelets-test`, which should trigger the [custom command](#custom-commands) to initiate unit testing
 
-### Installing `shapelets` on WSL (Windows Subsystem for Linux)
+### Mac OS
 
-For Windows users, we recommend using WSL (Windows Subsystem for Linux) which can easily be installed through the Microsoft Store. Please install Ubuntu 22.04 LTS. Additionally, Python 3.7+ should come pre-installed with WSL. Check this via `python3 --version`.
+1. Install [Homebrew](https://brew.sh/); it is an amazing package manager that will make this process easier
+2. Install `git`, `python3`, and `pip` via `brew install python3` (which also installs pip) then `brew install git`
+3. Clone the `shapelets` repository via `git clone https://github.com/uw-comphys/shapelets.git`
+4. To install `shapelets`, navigate to your /shapelets directory, then type `pip3 install .`
+5. To ensure correct installation, type `shapelets-test`, which should trigger the [custom command](#custom-commands) to initiate unit testing 
 
-To be able to use the [custom commands](#custom-commands) developed for `shapelets` (also known as entry points), please perform the following steps.
+### Windows
 
-* In your WSL terminal, type `nano ~/.bashrc`
-* At the bottom of the file, append the following line `export PATH="/home/user/.local/bin:$PATH"` where `user` is the username of your WSL unix profile
-* Press `CTRL+S` then `CTRL+X` to exit the editor
[…]
