### 2023-10-04 8197ee5
updated the example_4 config general section with fits_name and associated files (#10)
- updated the example_4 config general section with fits_name entry to keep config structure consistent with self-assembly method
- updated run.py to search for .fits files in the general section of the config
- update example_4.py to use new variable names (such as fits_name)
- fixed issue where .rst used incorrect image file types
