+++ examples/example_4/example_4.py
@@ -25,7 +25,7 @@ from shapelets.astronomy import (
 )
 
 ## Section 2: parameters
-fits_path = "images/galaxies.fits"
+fits_name = "galaxies.fits"
 shapelet_order = 10
 compression_order = 20
 
@@ -34,11 +34,13 @@ compression_order = 20
 # 3.1: loading .fits data and output directory handling
 save_path = os.getcwd()+'/output/'
 if not os.path.exists(save_path): os.mkdir("output")
+fits_path = os.getcwd()+'/images/' + fits_name
+
 output_base_path = save_path+fits_path[fits_path.rfind('/'):-5]
-data = load_fits_data(fits_path)
+fits_data = load_fits_data(fits_path)
 
 # 3.2 iidentifying areas in the image that contain decomposable galaxies
-(galaxy_stamps, star_stamps, noiseless_data) = get_postage_stamps(data, output_base_path)
+(galaxy_stamps, star_stamps, noiseless_data) = get_postage_stamps(fits_data, output_base_path)
 
 # 3.3 starting with the biggest galaxy, decomposes subdomain into a collection of shapelet coefficients
 decompose_galaxies(galaxy_stamps, star_stamps, noiseless_data, shapelet_order, compression_order, output_base_path)
