### 2023-10-08 bfbd3a8
Installation instructions updated for macOS
