### 2023-10-09 #5911 @nasserma
Happy (Canadian) Thanksgiving! I just wanted to check-in on the status of assigning an editor and also point out a somewhat special situation regarding the development of the [shapelets](https://github.com/uw-comphys/shapelets) package.

The core package was developed internally and included as [supplemental information](https://journals.aps.org/pre/supplemental/10.1103/PhysRevE.91.033307) for the research publication that introduced one of the application methods (self_assembly). This code was developed by R. Suderman   (@rsuderman) and later modified by T. Akdeniz as part of another research publication; both are co-authors of this submission.

Myself, @mptino, and @abaacus have rewritten, enhanced, and added to the core package in order to improve its usability for researchers interested in the use of shapelet functions for their own purposes, but also included several past applications (astronomy and self_assembly). Thus the three of us have made all of the commits to the relatively new public repo, the past contributions of Rob and Thomas are in various internal repos and are not reflected in the public repo. However,  all of the authors have made significant contributions to both the development of methods and the code itself.
