### 2023-10-10 #5911 @kyleniemeyer
Hi @AoifeHughes, could you edit this submission? It may fall outside your primary areas of expertise but involves image analysis. Thanks!
