### 2023-10-18 #5911 @xuanxu
The pdf compilation is failing because the metadata of the `paper.md` file includes [incorrect](https://github.com/uw-comphys/shapelets/blob/publications/publications/2309_JOSS/paper.md?plain=1#L15) [values](https://github.com/uw-comphys/shapelets/blob/publications/publications/2309_JOSS/paper.md?plain=1#L21) for some authors' orcid.
Updating those `XXXX-XXXX-XXXX-XXXX` values or removing them should fix it.
