### 2023-10-18 #5911 @danielskatz
@xuanxu - we used to get error messages in this case - can we get them back? (It will save you time too 🙂)
