### 2023-10-19 #5911 @nasserma
The invalid ORCIDs and missing DOIs have been corrected.
