### 2023-10-19 327ea76
Removed invalid ORCIDs and added missing DOIs to JOSS manuscript. (#12)
