+++ publications/2309_JOSS/paper.bib
@@ -17,7 +17,8 @@
   number={1},
   pages={197--210},
   year={2005},
-  publisher={Blackwell Science Ltd Oxford, UK}
+  publisher={Blackwell Science Ltd Oxford, UK},
+  doi={10.1111/j.1365-2966.2005.09453.x}
 }
 
 @article{Tino2023,
@@ -39,7 +40,8 @@
   number={7},
   pages={075703},
   year={2018},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/1361-6528/aaf353}
 }
 
 @article{Suderman2015,
@@ -50,7 +52,8 @@
   number={3},
   pages={033307},
   year={2015},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physreve.91.033307}
 }
 
 @article{Berge2019,
@@ -61,7 +64,8 @@
   number={1},
   pages={544--559},
   year={2019},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/mnras/stz787}
 }
 
 @article{Abukhdeir2016,
@@ -72,7 +76,8 @@
   number={8},
   pages={082001},
   year={2016},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/2053-1591/3/8/082001}
 }
 
 @article{scipy,
@@ -108,7 +113,8 @@
   number={23},
   pages={5760--5769},
   year={2004},
-  publisher={Optica Publishing Group}
+  publisher={Optica Publishing Group},
+  doi={10.1364/opex.12.005760}
 }
 
 @article{Sharpee2009,
@@ -118,7 +124,8 @@
   volume={26},
   pages={203--218},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10827-008-0107-5}
 }
 
 @article{Victor2006,
@@ -129,7 +136,8 @@
   number={1},
   pages={379--400},
   year={2006},
-  publisher={American Physiological Society}
+  publisher={American Physiological Society},
+  doi={10.1152/jn.00498.2005}
 }
 
 @article{Birrer2015,
@@ -140,7 +148,8 @@
   number={2},
   pages={102},
   year={2015},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/0004-637x/813/2/102}
 }
 
 @article{Lentati2015,
@@ -151,7 +160,8 @@
   number={3},
   pages={2159--2168},
   year={2015},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/mnras/stu2611}
 }
 
 @article{Desvignes2016,
@@ -178,5 +188,6 @@
   booktitle={Bond-Orientational Order in Condensed Matter Systems},
   pages={1--31},
   year={1992},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-1-4612-2812-7_1}
 }
