### 2023-11-13 #6058 @editorialbot
[REVIEW]: Shapelets: A Python package implementing shapelet functions and their applications
**Submitting author:** @nasserma (<a href="http://orcid.org/0000-0002-1772-0376">Nasser Mohieddin Abukhdeir</a>)
**Repository:** https://github.com/uw-comphys/shapelets
**Branch with paper.md** (empty if default branch): publications
**Version:** v1.0
