### 2023-11-16 #6058 @nasserma
> ```
> Software report:
>
> github.com/AlDanial/cloc v 1.88  T=0.02 s (708.9 files/s, 72487.7 lines/s)
> -------------------------------------------------------------------------------
> Language                     files          blank        comment           code
> -------------------------------------------------------------------------------
> Markdown                         3            131              0            220
> TeX                              1             16              0            177
> Python                           6            138            350            173
> YAML                             1              1              0             18
> TOML                             1              0              0              3
> -------------------------------------------------------------------------------
> SUM:                            12            286            350            591
> -------------------------------------------------------------------------------
>
>
> gitinspector failed to run statistical information for the repository
> ```

@AoifeHughes Could this software report be rerun, I am relatively sure it was using the publications branch of the repository, which did not have up-to-date code, I just merged it with the main branch and if the report is run now it will reflect the actual number of source files, lines of code, comments, etc.
