### 2023-11-21 0807f30
Added fix for WSLRegisterDistribution error for Windows installation instructions (#15)
* Added fix for WSLRegisterDistribution error for Windows installation instructions

* Updated command to invoke unit testing in entry_points.py
