### 2023-12-06 #6058 @tbmiller-astro
Hi, I was actually able to get my initial review done this afternoon. The package overall looks great, well done @nasserma and the other authors.

There are a few issues I found. Two major ones to do with missing `tests/` folder and need for more API documentation that I created issues for on the package's GitHub page.
