### 2023-12-06 #6058 @mptino
Hi @tbmiller-astro - thanks for the comments. I closed one issue regarding the unit testing and `shapelets-test` entry point (which is now functional) and uploaded a set of minimal tests for basic functionality of the package. More unit tests will appear in the coming weeks.
