### 2023-12-07 bd698bf
First unit tests for shapelet functions and self_assembly sub-module (#19)
