### 2023-12-27 88d733c
Many changes to package README and other documentation
