### 2023-12-27 028c5ff
Many changes to package README and other documentation (#21)
* Many changes to package README and other documentation

* Minor package README changes
