### 2024-01-09 6fcfa61
Update docs.yml
Corrected custom make for pdocs
