+++ .github/workflows/docs.yml
@@ -27,7 +27,7 @@ jobs:
       - run: pip install -e .
       # ADJUST THIS: build your documentation into docs/.
       # We use a custom build script for pdoc itself, ideally you just run `pdoc -o docs/ ...` here.
-      - run: python docs/make.py
+      - run: pdoc -o docs/ shapelets
 
       - uses: actions/upload-pages-artifact@v3
         with:
