### 2024-01-09 563e56c
Added self_assembly unit tests and extended invalid input checks
