+++ shapelets/self_assembly/quant.py
@@ -34,22 +34,27 @@ __all__ = [
     'rdistance'
 ]
 
-def convresponse(image, l, shapelet_order = 'default', normresponse = 'Vector'):
+def convresponse(image: np.ndarray, l: float, shapelet_order = 'default', 
+                 normresponse: str = 'Vector', verbose: bool = True):
     r""" This function computes the convolution between a range of 
-         shapelets and an image.
+         shapelets kernels and an image, extracting the magnitude of response
+         as well as the shapelet-based orientation.
     
     Parameters
     ----------
     image : numpy.ndarray
-        The image for convolution with shapelet functions.
+        The image to be convolved with shapelet kernels.
     l : float
         The characteristic wavelength of the image.
     shapelet_order : str or int
         'default' to use higher-order shapelets (m<=m') as in ref [3].
-        Can also accept integer value to use shapelets for m <= int.
+        Can also accept integer value to use shapelets for m = [1, shapelet_order].
     normresponse : str, optional
         Normalize magnitude of response (omega) in terms of response vectors = "Vector". Default. 
         Normalize each m-fold response wrt itself on [0, 1) = "Individual".
+    verbose: bool
+        True (default) to print out results of orientation algorithm to console.
+        False to not print any results or information to console.
 
     Returns
     -------
@@ -69,9 +74,27 @@ def convresponse(image, l, shapelet_order = 'default', normresponse = 'Vector'):
     .. [3] TODO: REFTINO
 
     """
-    if shapelet_order == 'default': mmax = None
-    else: mmax = int(shapelet_order)
+    if not isinstance(image, np.ndarray):
+        raise TypeError('image parameter must be a numpy array.')
+
+    if isinstance(shapelet_order, str):
+        if shapelet_order == 'default': 
+            mmax = None
+        else:
+            raise ValueError('shapelet_order parameter as a str must be "default".')
+        
+    elif isinstance(shapelet_order, int):
+        mmax = shapelet_order
+
+    else:
+        raise TypeError('shapelet_order parameter must either be "default" or integer value.')
     
+    if not isinstance(normresponse, str):
+        raise TypeError('normresponse parameter must be of type str.')
+    
+    elif normresponse not in ['Vector', 'Individual']:
+        raise ValueError('Valid normresponse parameters are "Vector" or "Individual".')
+
     minRespTol = 0.1
     
     Ny, Nx = image.shape
@@ -106,7 +129,8 @@ def convresponse(image, l, shapelet_order = 'default', normresponse = 'Vector'):
             con = fftconvolve(image, shapelet, mode = 'same')
             omega[:,:,i] = np.abs(con)
             phi[:,:,i] = np.angle(con)
-        print(f"Convolution complete for shapelets m <= {mmax}")
+        if verbose:
+            print(f"Convolution complete for shapelets m <= {mmax}")
 
     else:
         omega = np.empty((Ny, Nx, 200)) 
@@ -149,7 +173,8 @@ def convresponse(image, l, shapelet_order = 'default', normresponse = 'Vector'):
         m -= 1
         omega = omega[:,:,:m]
         phi = phi[:,:,:m]
-        print(f"Convolution complete for shapelets m <= {m} before tolerance exceeded")
+        if verbose:
+            print(f"Convolution complete for shapelets m <= {m} before tolerance exceeded")
 
     if normresponse == 'Vector': 
         norms = np.linalg.norm(omega, axis = 2)
@@ -167,20 +192,23 @@ def convresponse(image, l, shapelet_order = 'default', normresponse = 'Vector'):
     return omega, phi
 
 
-def defectid(response, l, pattern_order, num_clusters):
+def defectid(response: np.ndarray, l: float, pattern_order: str,
+             num_clusters: int or str):
     r""" Defect identification method from [1].
 
     Parameters
     ----------
[…]
