### 2024-01-10 f2a0e32
Re-organized run.py and added exceptions for incorrect config file usage
