+++ shapelets/run.py
@@ -19,11 +19,11 @@ import ast
 import configparser
 import os
 
+from .astronomy.galaxy import *
+
 from .self_assembly.misc import *
 from .self_assembly.quant import *
 from .self_assembly.wavelength import *
-
-from .astronomy.galaxy import *
 
 def run(config_file: str) -> None:
     r""" 
@@ -32,12 +32,18 @@ def run(config_file: str) -> None:
     
     Parameters
     ----------
-
     config_file : str
         The name of the config file.
-    """
+    
+    Notes
+    -----
+    Differentiation between submodule use is based on image_name or fits_name provided in config file
+        ...this may need to be changed in the future if more astronomy functionality is added.
 
-    # instance and read
+    """
+    ## configparser setup and input/output path organization ##
+
+    # instantiate and read
     config = configparser.ConfigParser()
     config.read(config_file)
 
@@ -52,7 +58,9 @@ def run(config_file: str) -> None:
     if not os.path.exists(save_path): 
         os.mkdir("output")
 
-    # parsing image
+    ## self_assembly submodule use ##
+        
+    # parsing input image of nanostructure
     if config.get('general', 'image_name', fallback=None):
         image_name = config.get('general', 'image_name')
         image = read_image(image_name = image_name, image_path = image_path)
@@ -60,71 +68,72 @@ def run(config_file: str) -> None:
         # obtain characteristic wavelength, needed for all self-assembly applications
         char_wavelength = get_wavelength(image = image)
 
+        if method == 'response_distance':
+            shapelet_order = config.get('response_distance', 'shapelet_order', fallback = 'default')
+            if shapelet_order != 'default':
+                shapelet_order = ast.literal_eval(shapelet_order)
+
+            num_clusters = config.get('response_distance', 'num_clusters', fallback = 'default')
+            if num_clusters != 'default':
+                num_clusters = ast.literal_eval(num_clusters)
+
+            ux = config.get('response_distance', 'ux', fallback = 'default')
+            uy = config.get('response_distance', 'uy', fallback = 'default')
+            if ux != 'default':
+                ux = ast.literal_eval(ux)
+            if uy != 'default':
+                uy = ast.literal_eval(uy)
+
+            response = convresponse(image = image, l = char_wavelength, shapelet_order = shapelet_order, normresponse = 'Vector')[0]
+            rd_field = rdistance(image = image, response = response, num_clusters = num_clusters, ux = ux, uy = uy)
+            process_output(image = image, image_name = image_name, save_path = save_path, output_from = 'response_distance',
+                           d = rd_field, num_clusters = num_clusters)
+
+        elif method == 'orientation':
+            pattern_order = config.get('orientation', 'pattern_order')
+
+            response, orients = convresponse(image = image, l = char_wavelength, shapelet_order = 6, normresponse = 'Individual')
+            mask, dilate, blended, maxval = orientation(pattern_order = pattern_order, l = char_wavelength, response = response, orients = orients)
+            process_output(image = image, image_name = image_name, save_path = save_path, output_from = 'orientation',
+                           mask = mask, dilate = dilate, orientation = blended, maxval = maxval)
+    
+        elif method == 'identify_defects':
+            pattern_order = config.get('identify_defects', 'pattern_order')
+
+            num_clusters = config.get('identify_defects', 'num_clusters', fallback = 'default') 
+            if num_clusters != 'default':
+                num_clusters = ast.literal_eval(num_clusters)
+
+            response = convresponse(image = image, l = char_wavelength, shapelet_order = 'default', normresponse = 'Vector')[0]
+            centroids, clusterMembers, defects = defectid(response = response, l = char_wavelength,
[…]
