### 2024-01-10 c56346d
Fixed error in defectid() radar chart output and other minor changes
