+++ shapelets/self_assembly/misc.py
@@ -29,11 +29,12 @@ __all__ = [
     'make_grid',
     'read_image',
     'process_output',
-    'trimage'
+    'trim_image'
 ]
 
 def make_grid(N: int):
-    r""" Make discretized grid based on width (N).
+    r""" 
+    Make discretized grid based on width (N).
     
     Parameters
     ----------
@@ -67,7 +68,8 @@ def make_grid(N: int):
 
 
 def read_image(image_name, image_path, verbose = True):
-    r""" Read an image using cv2 (OpenCV) with some extra handling.
+    r""" 
+    Read an image using cv2 (OpenCV) with some extra handling.
     
     Parameters
     ----------
@@ -114,10 +116,13 @@ def read_image(image_name, image_path, verbose = True):
 
 
 def process_output(image, image_name, save_path, output_from, **kwargs) -> None:
-    r""" Processes and saves output from any of the functions below,
+    r""" 
+    Processes and saves output from any of the functions below,
         * shapelets.self_assembly.quant.rdistance()
         * shapelets.self_assembly.quant.orientation()
         * shapelets.self_assembly.quant.defectid()
+    
+    It was used to generate Figures 6, 7, 8, and 9 from [1].
 
     Parameters
     ----------
@@ -136,6 +141,10 @@ def process_output(image, image_name, save_path, output_from, **kwargs) -> None:
     * output_from = 'orientation'         -->     mask, dilate, orientation, maxval
     * output_from = 'identify_defects'    -->     defects, centroids, clusterMembers
 
+    References
+    ----------
+    .. [1] TODO: REFTINO
+
     Examples
     --------
     
@@ -152,7 +161,7 @@ def process_output(image, image_name, save_path, output_from, **kwargs) -> None:
         # final image processing for response distance scalar field
         d = (d-d.min()) / (d.max()-d.min())
         d = 1-d
-        d = trimage(im=d, l=char_wavelength)
+        d = trim_image(im=d, l=char_wavelength)
 
         # plot and save
         plt.figure()
@@ -162,7 +171,7 @@ def process_output(image, image_name, save_path, output_from, **kwargs) -> None:
         plt.savefig(fname=plotname1, bbox_inches='tight', pad_inches=0)
         
         plt.figure()
-        plt.imshow(trimage(image,char_wavelength), cmap='gray')
+        plt.imshow(trim_image(image,char_wavelength), cmap='gray')
         plt.imshow(d,alpha=0.7, cmap='summer')
         plt.axis('off')
         plotname2 = f"{image_name[:-4]}_response_distance_overlay_k{num_clusters}.png"
@@ -178,7 +187,7 @@ def process_output(image, image_name, save_path, output_from, **kwargs) -> None:
         
         # plot and save
         plt.figure()
-        plt.imshow(trimage(image,char_wavelength), cmap='gray', alpha = 0.5)
+        plt.imshow(trim_image(image,char_wavelength), cmap='gray', alpha = 0.5)
         mask[mask == 0.0] = np.nan
         plt.imshow(mask, cmap='hsv', vmin=0, vmax=maxval)
         plt.axis('off')
@@ -186,7 +195,7 @@ def process_output(image, image_name, save_path, output_from, **kwargs) -> None:
         plt.savefig(fname=plotname1, dpi=600, bbox_inches='tight', pad_inches=0)
         
         plt.figure()
-        plt.imshow(trimage(image,char_wavelength), cmap='gray')
+        plt.imshow(trim_image(image,char_wavelength), cmap='gray')
         plt.imshow(dilate, cmap='hsv', alpha = 0.7, vmin=0, vmax=maxval)
         plt.axis('off')
         plotname2 = f"{image_name[:-4]}_orientation_dilate.png"
[…]
