### 2024-01-10 e08ff41
Unified style and cleaned up documentation across all sub-modules
