+++ shapelets/astronomy/misc.py
@@ -23,9 +23,10 @@ from ..functions import cartesian1D, cartesian2D
 
 SQRTPI = np.sqrt(np.pi)
 
-# Shapelet Methods ---------------------------------------------------------
 def decompose_analytic(image: np.ndarray, n_max: int, beta: float, centroid: np.ndarray, nspace: list[tuple[int, int]]=None) -> np.ndarray:
-    r"""Decomposes image into shapelet coefficents analytically by evaluating a series of integrals
+    r"""
+    Decomposes image into shapelet coefficents analytically by evaluating a 
+    series of integrals.
 
     Parameters
     ----------
@@ -47,7 +48,8 @@ def decompose_analytic(image: np.ndarray, n_max: int, beta: float, centroid: np.
     
     Notes
     -----
-    EXPANDS f INTO SMALLEST SQUARE MATRIX THAT CAN CONTAIN f
+    Expands f into smallest square matrix that can contain f.
+
     """
     if nspace == None:
         nspace = get_nspace(n_max)
@@ -77,7 +79,8 @@ def decompose_analytic(image: np.ndarray, n_max: int, beta: float, centroid: np.
     return coefficients
 
 def decompose_kernel(image: np.ndarray, n_max: int, beta: float, centroid: np.ndarray, nspace: list[tuple[int, int]]=None) -> np.ndarray:
-    r"""Decomposes image into shapelet coefficents using projection of image onto kernels
+    r"""
+    Decomposes image into shapelet coefficents using projection of image onto kernels.
 
     Parameters
     ----------
@@ -96,9 +99,6 @@ def decompose_kernel(image: np.ndarray, n_max: int, beta: float, centroid: np.nd
     -------
     coefficients : np.ndarray (dtype=float)
         matrix of shapelet coefficients representing the decomposition of the image
-    
-    Notes
-    -----
 
     """
     coefficients = np.zeros([n_max+1]*2)
@@ -120,7 +120,8 @@ def decompose_kernel(image: np.ndarray, n_max: int, beta: float, centroid: np.nd
     return coefficients
 
 def reconstruct(s_coeff: np.ndarray, n_max: int, beta: float, centroid: np.ndarray, dimensions: tuple[int, int], nspace: list[tuple]=None) -> np.ndarray:
-    r"""Given a matrix of shapelet coefficients, constructs image
+    r"""
+    Given a matrix of shapelet coefficients, constructs image.
 
     Parameters
     ----------
@@ -142,11 +143,7 @@ def reconstruct(s_coeff: np.ndarray, n_max: int, beta: float, centroid: np.ndarr
     image : np.ndarray (dtype=float)
         the reconstructed image
     
-    Notes
-    -----
-
     """
-
     if nspace == None:
         nspace = get_nspace(n_max)
 
@@ -164,8 +161,9 @@ def reconstruct(s_coeff: np.ndarray, n_max: int, beta: float, centroid: np.ndarr
 
 # calculate improved parameters to optimize shapelet coeffcients
 def update_shapelet_parameters(coeff: np.ndarray, n_max: int, beta: float, centroid: np.ndarray, nspace: list[tuple]=None) -> tuple[float, np.ndarray]:
-    r"""Calculates an object's characteristic scale and centroid using its calculated shapelet coefficients.
-    These new parameters can be used to decompose the image again with lower error
+    r"""
+    Calculates an object's characteristic scale and centroid using its calculated shapelet coefficients.
+    These new parameters can be used to decompose the image again with lower error.
 
     Parameters
     ----------
[…]
