+++ shapelets/self_assembly/misc.py
@@ -50,9 +50,7 @@ def make_grid(N: int):
     
     Notes
     -----
-    As per convention, N should only be an odd number.
-    Additionally, note that grid_x == grid_y (identical for obvious
-        reasons).
+    As per convention, N should only be an odd number. Additionally, note that grid_x == grid_y (identical for obvious reasons).
 
     """
     if N % 2 == 0:
@@ -86,8 +84,7 @@ def read_image(image_name: str, image_path: str, verbose: bool = True):
 
     Notes
     -----
-    The bounds for either threshold (-1, 1) are intentional to align with 
-    the minimum and maximum of shapelet function intensity.
+    The bounds for either threshold (-1, 1) are intentional to align with the minimum and maximum of shapelet function intensity.
     
     """
     if os.path.exists(image_path):
@@ -110,17 +107,13 @@ def read_image(image_name: str, image_path: str, verbose: bool = True):
 def process_output(image: np.ndarray, image_name: str, save_path: str, output_from: str, **kwargs) -> None:
     r""" 
     Processes and saves output from any of the functions below,
-        * shapelets.self_assembly.quant.rdistance()
-        * shapelets.self_assembly.quant.orientation()
-        * shapelets.self_assembly.quant.defectid()
+    * shapelets.self_assembly.quant.rdistance()
+    * shapelets.self_assembly.quant.orientation()
+    * shapelets.self_assembly.quant.defectid()
     
     It was used to generate Figures 6, 7, 8, and 9 from [1].
 
-    NOTE: any image saved from the **kwargs argument is trimmed using 
-        shapelets.self_assembly.misc.trim_image().
-        This is because the convolution with shapelet kernels is padded on
-        the edges, producing a fuzzy convolutional response.
-        The trim_image() function removes this fuzzy response.
+    NOTE: any image saved from the **kwargs argument is trimmed using shapelets.self_assembly.misc.trim_image(). This is because the convolution with shapelet kernels is padded on the edges, producing a fuzzy convolutional response. The trim_image() function removes this fuzzy response.
 
     Parameters
     ----------
@@ -131,8 +124,7 @@ def process_output(image: np.ndarray, image_name: str, save_path: str, output_fr
     save_path : str
         The path to save results.
     output_from : str
-        The name of the method for which we will process and save the output/results.
-        Options are: 'response_distance', 'orientation', or 'identify_defects'
+        The name of the method for which we will process and save the output/results. Options are: 'response_distance', 'orientation', or 'identify_defects'
 
     Notes
     -----
@@ -306,8 +298,7 @@ def process_output(image: np.ndarray, image_name: str, save_path: str, output_fr
 
 def image_difference(im1: np.ndarray, im2: np.ndarray):
     r""" 
-    This function computes the normalized difference between two images.
-    It was used to generate Figure 5 from [1].
+    This function computes the normalized difference between two images. It was used to generate Figure 5 from [1].
 
     Parameters
     ----------
@@ -344,8 +335,7 @@ def image_difference(im1: np.ndarray, im2: np.ndarray):
 
 def trim_image(im: np.ndarray, l: float):
     r""" 
-    Trim image edges based on characteristic wavelength (l). 
-    Useful for images post convolution, as edges can present distortions.
+    Trim image edges based on characteristic wavelength (l). Useful for images post convolution, as edges can present distortions because of padded convolution.
 
     Parameters
     ----------
@@ -360,8 +350,7 @@ def trim_image(im: np.ndarray, l: float):
 
     Notes
     -----
-    The characteristic wavelength is roughly the distance between feature centers,
-    thus making it an appropriate size for image trim or truncation after convolution.
+    The characteristic wavelength is roughly the distance between feature centers, thus making it an appropriate size for image trim or truncation after convolution. 
 
     """
     trim = int(l)
