### 2024-01-10 10c73ff
Converted docstring of all code from .rst to .md
