### 2024-01-11 c7cd244
More fixes and changes to markdown function docstrings
