+++ shapelets/functions.py
@@ -29,29 +29,29 @@ __all__ = [
     'exponential2D'
 ]
 
-def cartesian1D(n: int, x1: Union[float,np.ndarray], beta: float = 1.):
+def cartesian1D(n: int, x1: Union[float,np.ndarray], beta: float = 1.) -> Union[float,np.ndarray]:
     r""" 
-    1D cartesian shapelet function defined as [1]_,
+    1D cartesian shapelet function defined as[1]_,
 
     $$ S_{n}(x; \beta) = \beta^{-\frac{1}{2}}  \phi_{n}(\frac{x}{\beta}) $$
 
     with $$ \phi_n(x) = \left( 2^n \pi^{\frac{1}{2}} n! \right)^{-\frac{1}{2}} H_n(x) exp(-\frac{x^2}{2}) $$
 
-    where $\phi_n$ is the dimensionless [shapelet] basis function, $\beta$ is the characteristic shapelet scale, $H_n$ is a hermite polynomial of order n, and $n$ is the shapelet order.
+    where $\phi_n$ is the dimensionless basis function, $\beta$ is the shapelet length scale, $H_n$ is a hermite polynomial of order $n$, and $n$ is the shapelet order parameter.
 
     Parameters
     ----------
     * n: int
-        * Shapelet order. Acceptable values $n \geq 0$.
-    * x1: float or np.ndarray
-        * The input to the shapelet function.
+        * Shapelet order. Acceptable values are $n \geq 0$
+    * x1: Union[float,np.ndarray]
+        * The input to the shapelet function
     * beta: float
-        * The characteristic shapelet length scale parameter.
+        * The shapelet length scale parameter
 
     Returns
     -------
-    * Sc(x1): float or np.ndarray
-        * Shapelet function evaluated at S(x1).
+    * Sc(x1): Union[float,np.ndarray]
+        * Shapelet function evaluated at (x1)
 
     References
     ----------
@@ -72,33 +72,33 @@ def cartesian1D(n: int, x1: Union[float,np.ndarray], beta: float = 1.):
 
     return Sc(x1)
 
-def cartesian2D(n1: int, n2: int, x1: Union[float,np.ndarray], x2: Union[float,np.ndarray], beta: float = 1.):
+def cartesian2D(n1: int, n2: int, x1: Union[float,np.ndarray], x2: Union[float,np.ndarray], beta: float = 1.) -> Union[float,np.ndarray]:
     r""" 
-    2D cartesian shapelet function defined as [1]_,
+    2D cartesian shapelet function defined as[1]_,
 
     $$ S_{n_1,n_2}(x_1, x_2; \beta) = \beta^{-1} \phi_{n_1}(\frac{x_1}{\beta}) \phi_{n_2}(\frac{x_2}{\beta}) $$
 
     with $$ \phi_n(x) = \left( 2^n \pi^{\frac{1}{2}} n! \right)^{-\frac{1}{2}} H_n(x) exp(-\frac{x^2}{2}) $$
 
-    where $\phi_n$ is the dimensionless [shapelet] basis function, $\beta$ is the characteristic shapelet scale, $H_n$ is a hermite polynomial of order n, and $n_1$ and $n_2$ are the shapelet orders.
+    where $\phi_n$ is the dimensionless basis function, $\beta$ is the shapelet length scale, $H_n$ is a hermite polynomial of order $n$, and $n_1$ and $n_2$ are the shapelet orders.
 
     Parameters
     ----------
     * n1: int
-        * Shapelet order in x direction. Acceptable values $n1 \geq 0$.
+        * Shapelet order in x direction. Acceptable values are $n1 \geq 0$
     * n2: int
[…]
