+++ shapelets/self_assembly/wavelength.py
@@ -26,19 +26,19 @@ __all__ = [
 
 def lambda_to_beta(m: int, l: float):
     r""" 
-    Converts lambda (l), the characteristic wavelength of the pattern [1]_ [2]_ to the appropriate beta value for orthonormal polar shapelets from ref. [3]_.
+    Converts lambda (l), the characteristic wavelength of the image[1]_ to the appropriate beta value for orthonormal polar shapelets[3]_ (see ``shapelets.functions.orthonormalpolar2D``).
     
     Parameters
     ----------
     * m: int
         * Shapelet degree of rotational symmetry
     * l: float
-        * The characteristic wavelength of the pattern
+        * The characteristic wavelength of the image[1]_
     
     Returns
     -------
     * beta: float
-        * The characteristic shapelet length scale parameter based on ref. [3]_
+        * The characteristic shapelet length scale parameter based on ref.[3]_
 
     References
     ----------
@@ -63,7 +63,7 @@ def lambda_to_beta(m: int, l: float):
 
 def get_wavelength(image: np.ndarray, rng: list = [0, 50], verbose: bool = True):
     r""" 
-    Find characteristic wavelength of an image. Computed from refs [1]_ [2]_.
+    Find characteristic wavelength of an image. Computed from refs[1,2]_.
 
     mpt@mpt: The rng needs to be optimized better. For example, if the wavelength is bigger than 50 need to fix this. But, if we just let rng = None (i.e., all wavelengths) then it will find way too big of a wavelength as the scale...
     
@@ -124,7 +124,7 @@ def get_wavelength(image: np.ndarray, rng: list = [0, 50], verbose: bool = True)
 
 def radialavg(image: np.ndarray):
     r""" 
-    Calculates the radially averaged intensity of an image. Based on work from [1]_ [2]_.
+    Calculates the radially averaged intensity of an image. Based on work from[1,2]_.
     
     Parameters
     ----------
