+++ shapelets/self_assembly/misc.py
@@ -29,6 +29,7 @@ __all__ = [
     'make_grid',
     'read_image',
     'process_output',
+    'image_difference',
     'trim_image'
 ]
 
@@ -50,11 +51,12 @@ def make_grid(N: int):
     
     Notes
     -----
-    As per convention, N should only be an odd number. Additionally, note that grid_x == grid_y (identical for obvious reasons).
+    As per convention, N should only be an odd number. Additionally, note that grid_x = grid_y.
 
     """
     if N % 2 == 0:
-        raise ValueError('N must be an odd number, not even.')
+        print('Detected even grid size, adding 1 to enforce odd rule See self_assembly.misc.make_grid() docs.')
+        N += 1
     if N < 3:
         raise ValueError('N must be at least 3 or greater.')
     
@@ -66,7 +68,7 @@ def make_grid(N: int):
 
 def read_image(image_name: str, image_path: str, verbose: bool = True):
     r""" 
-    Read an image using cv2 (OpenCV) with some extra handling.
+    Read an image using OpenCV, with some extra handling. By default, re-scales images as greyscale on [-1, 1].
     
     Parameters
     ----------
@@ -74,17 +76,17 @@ def read_image(image_name: str, image_path: str, verbose: bool = True):
         * The filename of the image (including extension)
     * image_path: str
         * The path holding the image
-    * verbose: bool
-        * True to print image-related information
+    * verbose: bool, optional
+        * True (default) to print image-related information
     
     Returns
     -------
     * f: np.ndarray
-        * The image as a numpy array.
+        * The image as a numpy.ndarray.
 
     Notes
     -----
-    The bounds for either threshold (-1, 1) are intentional to align with the minimum and maximum of shapelet function intensity.
+    Re-scaling of image to greyscale on [-1, 1] is intentional to align with the minimum and maximum of shapelet function values.
     
     """
     if os.path.exists(image_path):
@@ -107,13 +109,13 @@ def read_image(image_name: str, image_path: str, verbose: bool = True):
 def process_output(image: np.ndarray, image_name: str, save_path: str, output_from: str, **kwargs) -> None:
     r""" 
     Processes and saves output from any of the functions below,
-    * shapelets.self_assembly.quant.rdistance()
-    * shapelets.self_assembly.quant.orientation()
-    * shapelets.self_assembly.quant.defectid()
+    * ``shapelets.self_assembly.quant.rdistance``
+    * ``shapelets.self_assembly.quant.orientation``
+    * ``shapelets.self_assembly.quant.defectid``
     
     It was used to generate Figures 6, 7, 8, and 9 from ref [1]_.
 
-    NOTE: any image saved from the **kwargs argument is trimmed using shapelets.self_assembly.misc.trim_image(). This is because the convolution with shapelet kernels is padded on the edges, producing a fuzzy convolutional response. The trim_image() function removes this fuzzy response.
+    NOTE: any image saved from the **kwargs argument is trimmed using ``shapelets.self_assembly.misc.trim_image``. This is because the convolution with shapelet kernels is padded on the edges, producing a fuzzy convolutional response. The ``shapelets.self_assembly.misc.trim_image`` function removes this fuzzy response.
 
     Parameters
     ----------
@@ -129,9 +131,9 @@ def process_output(image: np.ndarray, image_name: str, save_path: str, output_fr
     Notes
     -----
     Required kwargs are,
-    * output_from = 'response_distance'   -->     d, num_clusters
-    * output_from = 'orientation'         -->     mask, dilate, orientation, maxval
-    * output_from = 'identify_defects'    -->     defects, centroids, clusterMembers
[…]
