### 2024-01-11 742a880
Final changes to function docstrings
