### 2024-01-11 42e3af2
Minor correction to astronomy function docstring
