### 2024-01-11 1879ca1
Converted all example documentation from .rst to .md
