### 2024-01-14 5d99f51
Updated documentation with latest shapelet paper doi
