### 2024-01-14 dce1e7f
Updated JOSS manuscript with accepted shapelet paper info
