+++ publications/2309_JOSS/paper.bib
@@ -21,15 +21,16 @@
   doi={10.1111/j.1365-2966.2005.09453.x}
 }
 
-@article{Tino2023,
+@article{Tino2024,
   title={Shapelet-based orientation and defect identification method for nanostructured surface imaging},
   author={Tino, Matthew Peres and Suderman, Robert and Abukhdeir, Nasser Mohieddin},
-  journal={submitted to Nanotechnology},
+  journal={accepted to Nanotechnology},
   volume={},
   number={},
   pages={},
-  year={2023},
-  publisher={IOP Publishing}
+  year={2024},
+  publisher={IOP Publishing},
+  doi={http://dx.doi.org/10.1088/1361-6528/ad1df4}
 }
 
 @article{Akdeniz2018,
