### 2024-01-18 d2073eb
Added astronomy unit tests and invalid input checks
- created unit tests for core functions in the astronomy submodule
- fixed error when the flag SHOW_STAMPS was set to False
