### 2024-01-19 c5a2b0b
Fixed unit test error and removed platform dependence from tests
