### 2024-01-19 23aebbb
Linked package documentation with website via pdoc
