+++ shapelets/docs/__init__.py
@@ -1,0 +1,35 @@
+########################################################################################################################
+# Copyright 2023 the authors (see AUTHORS file for full list).                                                         #
+#                                                                                                                      #
+# This file is part of shapelets.                                                                                      #
+#                                                                                                                      #
+# Shapelets is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General       #
+# Public License as published by the Free Software Foundation, either version 2.1 of the License, or (at your option)  #
+# any later version.                                                                                                   #
+#                                                                                                                      #
+# Shapelets is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied      #
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more  #
+# details.                                                                                                             #
+#                                                                                                                      #
+# You should have received a copy of the GNU Lesser General Public License along with shapelets. If not, see           #
+# <https://www.gnu.org/licenses/>.                                                                                     #
+########################################################################################################################
+
+r"""
+
+# Package Documentation
+
+## [Installation](https://mptino.github.io/shapelets/shapelets/docs/installation.html)
+
+## [Examples](https://mptino.github.io/shapelets/shapelets/docs/examples.html)
+
+## [Custom Commands & Methods of Usage](https://mptino.github.io/shapelets/shapelets/docs/usage.html)
+
+## [Additional WSL Instructions](https://mptino.github.io/shapelets/shapelets/docs/WSL.html)
+
+"""
+
+from .examples import *
+from .installation import *
+from .usage import *
+from .WSL import *
