### 2024-01-19 2f57a2f
Fixed website documentation errors and rescaled images
