+++ README.md
@@ -1,16 +1,6 @@
-## What is `shapelets`?
+# What is shapelets? 
 
-[**Summary**](#summary)
-| [**Installation**](#installation)
-| [**Examples**](#examples-of-usage)
-| [**Contribute**](#contribute)
-| [**Issues**](#issues)
-| [**Authors**](#authors)
-
-
-## Summary 
-
-`shapelets` is a python package that implements several shapelet functions and some of their significant applications in science and astronomy. These functions form a complete and orthonormal set, allowing them to capture complex geometries and information from any physical shape. Furthermore, shapelets are localized and can be scaled to match that of any physical feature. 
+shapelets is a python package that implements several shapelet functions and some of their significant applications in science and astronomy. These functions form a complete and orthonormal set, allowing them to capture complex geometries and information from any physical shape. Furthermore, shapelets are localized and can be scaled to match that of any physical feature. 
 
 Due to these properties, they have seen extensive use in recent years, with several different formulations and applications:
 
@@ -19,107 +9,31 @@ Due to these properties, they have seen extensive use in recent years, with seve
 * Computational neuroscience ([J. D. Victor (2006)](https://doi.org/10.1152/jn.00498.2005), [T. O. Sharpee (2009)](https://doi.org/10.1007%2Fs10827-008-0107-5))
 * Medical imaging ([J. Weissman (2004)](https://doi.org/10.1364/OPEX.12.005760))
 
-The `shapelets` package provides reference documentation and code for four (4) shapelet functions: 
+The shapelets package provides reference documentation and code for four (4) shapelet functions: 
 
 * Cartesian shapelets ([A. Refregier (2003)](https://doi.org/10.1046/j.1365-8711.2003.05901.x)), 
 * Polar shapelets ([R. Massey (2005)](https://doi.org/10.48550/arXiv.astro-ph/0408445)),
 * Orthonormal polar shapelets with constant radial scale ([T. Akdeniz (2018)](https://doi.org/10.1088/1361-6528/aaf353)), and 
 * Exponential shapelets ([J. Berge (2019)](https://doi.org/10.48550/arXiv.1903.05837))
 
+# Getting Started
 
-## Installation
+Users are directed to the shapelets [official website](https://mptino.github.io/shapelets/shapelets.html) for all documentation, including installation instructions, detailed examples, and more. 
 
-`shapelets` requires Python 3.10 or greater.
-Installing `shapelets` is highly dependent on the operating system of your machine. 
+# Contribute
 
-### Mac OS
-
-Please ensure your macOS is at least macOS 12 Monterey (released 2021).
-
-1. Install [Homebrew](https://brew.sh/) by copy and pasting the command website homepage into your terminal
-2. Install `git`, `python3.10`, and `pip` in the terminal: `brew install git python@3.10` (automatically installs pip) 
-3. Clone this repository: `git clone https://github.com/uw-comphys/shapelets.git`
-4. Navigate to the top-most directory (i.e. ./shapelets) and install this package: `pip3.10 install .`
-5. Ensure correct installation by running unit tests via the [custom command](#custom-commands): `shapelets-test`
-
-### Windows
-
-The instructions for Windows use WSL (Windows Subsystem for Linux). 
-
-1. Install Ubuntu 22.04 LTS from the Microsoft Store 
-2. Open Ubuntu 22.04 LTS and create a Unix profile (username and password) 
-3. If the following error occurs: `WSLRegisterDistribution failed with error: 0x80370102`, follow these steps
-	* Open *Turn Windows Features on and off*, and enable the following
-	* *Virtual Machine Platform*, *Windows Subsystem for Linux*, and *Hyper-V* (if available) 
[…]
