+++ shapelets/docs/usage.py
@@ -22,7 +22,7 @@ r"""
 The shapelets package makes use of [entry points](https://packaging.python.org/en/latest/specifications/entry-points/), which are custom command-line arguments that trigger specific tasks.
 
 * `shapelets CONFIG` - here, `CONFIG` is the name of the plaintext configuration file that will run specific package methods based on the parameters listed in the file
-* `shapelets-test` - triggers all unit tests in the [tests](https://github.com/uw-comphys/shapelets/tree/main/tests) directory and will report any failures. It is encouraged to use this command if modifying the package source code and after initial package [installation](https://mptino.github.io/shapelets/shapelets/docs/installation.html)
+* `shapelets-test` - triggers all unit tests in the [tests](https://github.com/uw-comphys/shapelets/tree/main/tests) directory and will report any failures. It is encouraged to use this command if modifying the package source code and after initial package [installation](https://uw-comphys.github.io/shapelets/shapelets/docs/installation.html)
 
 # Methods of Usage 
 
@@ -36,6 +36,6 @@ The shapelets package can be used in two different formats.
     from shapelets.functions import cartesian2D
     ```
 
-All [examples](https://mptino.github.io/shapelets/shapelets/docs/examples.html) provide implementation using both the configuration file method and scripting method.  
+All [examples](https://uw-comphys.github.io/shapelets/shapelets/docs/examples.html) provide implementation using both the configuration file method and scripting method.  
 
 """
