+++ shapelets/_run.py
@@ -25,8 +25,8 @@ from .self_assembly.misc import *
 from .self_assembly.quant import *
 from .self_assembly.wavelength import *
 
-def run(config_file: str) -> None:
-    r""" @private
+def _run(config_file: str) -> None:
+    r"""
     Main run function that handles input configuration file.
     
     Parameters
