### 2024-01-19 2c152e2
Finished website documentation and fixed unit test error (#28)
* Fixed unit test error and removed platform dependence from tests

* Linked package documentation with website via pdoc

* Fixed website documentation errors and rescaled images

* Multiple documentation changes and fixes

* Finished website documentation and made internal modules private
