+++ shapelets/docs/examples.py
@@ -1,0 +1,593 @@
+########################################################################################################################
+# Copyright 2023 the authors (see AUTHORS file for full list).                                                         #
+#                                                                                                                      #
+# This file is part of shapelets.                                                                                      #
+#                                                                                                                      #
+# Shapelets is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General       #
+# Public License as published by the Free Software Foundation, either version 2.1 of the License, or (at your option)  #
+# any later version.                                                                                                   #
+#                                                                                                                      #
+# Shapelets is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied      #
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more  #
+# details.                                                                                                             #
+#                                                                                                                      #
+# You should have received a copy of the GNU Lesser General Public License along with shapelets. If not, see           #
+# <https://www.gnu.org/licenses/>.                                                                                     #
+########################################################################################################################
+
+r"""
+
+# Examples Overview
+
+The example files and scripts can be found in the [github example directory](https://github.com/uw-comphys/shapelets/tree/main/examples).
+Associated documentation can be found below and should be explored by all new users. 
+
+
+## Example 1 - Response Distance
+
+This example goes through the process of computing the response distance method ([R. Suderman (2015)](https://doi.org/10.1103/PhysRevE.91.033307)) for self-assembly microscopy images using the ``shapelets.self_assembly`` submodule.
+
+The files for this example can be found [here](https://github.com/uw-comphys/shapelets/tree/main/examples/example_1).
+
+**NOTE** - this example can be run in two different ways, and both methods are presented here.
+* (1) the configuration-file based user interface [here](#config-method---config-setup)
+* (2) importing neccessary shapelets submodules and methods in a script-based format [here](#scripting-method---example_1py-breakdown)
+
+### Technical overview
+
+The response distance ([R. Suderman (2015)](https://doi.org/10.1103/PhysRevE.91.033307)) is calculated as:
+
+$$ d_{i, j} = \min \| \vec{R} - \vec{r_{i,j}} \|_2 $$
+
+where $\vec{r_{i,j}}$ denotes the given response vector at pixel location $\{i, j\}$ and $\vec{R}$ is the reference set (or subdomain) of response vectors.
+
+### Directory overview
+
+The example [directory](https://github.com/uw-comphys/shapelets/tree/main/examples/example_1) should contain the following.
+
+![](images/example_1_dir.png)
+
+* **config** contains the configuration file to run example 1 via config method
+* **example_1.py** contains the script to run example 1 via scripting method
+* **images/** contains the simulated stripe self-assembly microscopy image ([R. Suderman (2015)](https://doi.org/10.1103/PhysRevE.91.033307)) used in this example, shown below
+
+![](images/lamSIM1.png)
+
+### Config method - config setup
+
+The *general* section of the configuration file contains two parameters. 
+
+	[general]
+	image_name = lamSIM1.png
+	method = response_distance
+
+The "image_name" and "method" parameters are required.
+
[…]

[…]
