### 2024-01-19 b21a38f
Fixed errors in example documentation and minor README change
