### 2024-01-20 04e0947
Removed stale links and referenced package website in paper.md
