+++ publications/2309_JOSS/paper.md
@@ -73,7 +73,7 @@ The table below summarizes the specific shapelet applications implemented in thi
 | Orientation                     | Local pattern orientation for self-assembly microscopy imaging [@Tino2024] via `shapelets.self_assembly.orientation`                    |
 | Defect identification     | Defect identification method for self-assembly microscopy imaging [@Tino2024] via `shapelets.self_assembly.defectid`                    |
 
-More information, such as installation instructions and application-specific examples can be found in the package [README](https://github.com/uw-comphys/shapelets) file.
+More information, such as installation instructions and application-specific examples can be found on the package [website](https://uw-comphys.github.io/shapelets) file.
 
 # User Interface Methods
 
@@ -85,7 +85,7 @@ The text-based user interface for `shapelets` is centered around configuration f
 
 ![Sample directory hierarchy.\label{fig:sample_dir}](images/user_interface.png){ width=40% }
 
-The main (shapelets_example) directory contains the main configuration file (e.g., `config`) which is a plain-text file specifying specific parameters or methods to be used. Acceptable parameters and options for the configuration file can be found throughout the [examples](https://github.com/uw-comphys/shapelets/tree/main/examples), depending on the shapelet-based method being applied.
+The main (shapelets_example) directory contains the main configuration file (e.g., `config`) which is a plain-text file specifying specific parameters or methods to be used. Acceptable parameters and options for the configuration file can be found throughout the examples, depending on the shapelet-based method being applied.
 
 The `images/` subdirectory contains the data for image analysis and must be present. The `output/` subdirectory is created by the `shapelets` software and contains output data/images based on the analysis directed in the `config`.
 
@@ -93,14 +93,14 @@ Alternatively, the Python-based software interface of the `shapelets` package ca
 
 # Examples of Usage
 
-For each example included in the package, implementations using both the configuration file and programming-based interfaces are demonstrated. Several detailed image processing examples were developed that demonstrate the use and capabilities of the shapelets package for both astronomy and self-assembly related applications. See [here](https://github.com/uw-comphys/shapelets/tree/main/examples) and [here](https://github.com/uw-comphys/shapelets/tree/main/docs/examples) for examples and their associated documentation, respectively.
+For each example included in the package, implementations using both the configuration file and programming-based interfaces are demonstrated. Several detailed image processing examples were developed that demonstrate the use and capabilities of the shapelets package for both astronomy and self-assembly related applications. See the package [website](https://uw-comphys.github.io/shapelets) for more example details and documentation, and the package [code](https://github.com/uw-comphys/shapelets) for related files.
 
-Examples 1-3 demonstrate use of the `shapelets.self_assembly` module, with specific applications for the response distance method [@Suderman2015], local pattern orientation [@Tino2024], and defect identification method [@Tino2024].
+Examples 1-3 demonstrate use of the `shapelets.self_assembly` module, with specific applications for the response distance method [@Suderman2015], defect identification method [@Tino2024], and local pattern orientation [@Tino2024] (respectively).
 Example 4 demonstrates use of the `shapelets.astronomy` submodule for the decomposition and reconstruction of galactic images. All examples have instructions to use the `shapelets` package via configuration files or importing relevant submodules in pre-configured `.py` files (scripting). Examples 1, 2, and 4 are shown here.
 
 ## Example 1 - Response Distance Method
 
[…]
