### 2024-01-20 6741b09
Fixed errors in example documentation and minor README change (#29)
