+++ shapelets/docs/examples.py
@@ -30,8 +30,8 @@ This example goes through the process of computing the response distance method 
 The files for this example can be found [here](https://github.com/uw-comphys/shapelets/tree/main/examples/example_1).
 
 **NOTE** - this example can be run in two different ways, and both methods are presented here.
-* (1) the configuration-file based user interface [here](#config-method---config-setup)
-* (2) importing neccessary shapelets submodules and methods in a script-based format [here](#scripting-method---example_1py-breakdown)
+* (1) the configuration-file based user interface 
+* (2) importing neccessary shapelets submodules and methods in a script-based format
 
 ### Technical overview
 
@@ -73,7 +73,7 @@ The *response_distance* section of the configuration file contains four paramete
 	ux = [50, 80]
 	uy = [150, 180]
 		
-These parameters are explained in detail in the [next section](#method-parameters).
+These parameters are explained in detail in the next section.
 
 ### Method parameters
 
@@ -83,25 +83,25 @@ Note these parameters are the same if using the configuration-file based method 
 
 These parameters are explained below, note that *default* refers to default behaviour if the parameter is excluded.
 
-* **shapelet_order** 
+**shapelet_order** 
 
-	* int - integer, compute convolution for maximum shapelet order $m'$, i.e. $m \in [1, m']$
-	* default = $m'$ computed by the higher-order shapelet algorithm ([M.P. Tino (2024)](http://dx.doi.org/10.1088/1361-6528/ad1df4))
+* int - integer, compute convolution for maximum shapelet order $m'$, i.e. $m \in [1, m']$
+* default = $m'$ computed by the higher-order shapelet algorithm ([M.P. Tino (2024)](http://dx.doi.org/10.1088/1361-6528/ad1df4))
 
-* **num_clusters**
+**num_clusters**
 
-	* int - integer (including 0, which will not perform k-means clustering and use all response vectors in subdomain)
-	* default = 20 (see [T. Akdeniz (2018](https://doi.org/10.1088/1361-6528/aaf353]))
+* int - integer (including 0, which will not perform k-means clustering and use all response vectors in subdomain)
+* default = 20 (see [T. Akdeniz (2018](https://doi.org/10.1088/1361-6528/aaf353]))
 
-* **ux**
+**ux**
 
-	* list - I.e., [10, 20], this would represent the lower and upper bound for the user-defined subdomain in x-direction respectively
-	* default = 'default', user select x-bounds during runtime, see [here](#selecting-subdomain-bounds-during-runtime)
+* list - I.e., [10, 20], this would represent the lower and upper bound for the user-defined subdomain in x-direction respectively
+* default = 'default', user select x-bounds during runtime, see [here](#selecting-subdomain-bounds-during-runtime)
 
-* **uy** 
+**uy** 
 
-	* list - I.e., [30, 40], this would represent the lower and upper bound for the user-defined subdomain in y-direction respectively
-	* default = 'default', user select y-bounds during runtime, see [here](#selecting-subdomain-bounds-during-runtime)
+* list - I.e., [30, 40], this would represent the lower and upper bound for the user-defined subdomain in y-direction respectively
+* default = 'default', user select y-bounds during runtime, see [here](#selecting-subdomain-bounds-during-runtime)
 
 **NOTE**
 
@@ -131,14 +131,14 @@ This method is presented as an alternative to the configuration-file based user 
 The code breakdown is as follows,
 
 * Section 1: importing modules - imports the necessary modules from the shapelets package
-* Section 2: parameters - this contains the required parameters needed for the methods required to compute the response distance method (see [method parameters](#method-parameters) for details)
+* Section 2: parameters - this contains the required parameters needed for the methods required to compute the response distance method (see Method parameters section for details)
[…]
