### 2024-02-07 9e8656e
JOSS Manuscript corrections and updated Tino reference
