+++ publications/2309_JOSS/paper.bib
@@ -24,13 +24,13 @@
 @article{Tino2024,
   title={Shapelet-based orientation and defect identification method for nanostructured surface imaging},
   author={Tino, Matthew Peres and Suderman, Robert and Abukhdeir, Nasser Mohieddin},
-  journal={accepted to Nanotechnology},
-  volume={},
-  number={},
-  pages={},
+  journal={Nanotechnology},
+  volume={35},
+  number={16},
+  pages={165705},
   year={2024},
   publisher={IOP Publishing},
-  doi={http://dx.doi.org/10.1088/1361-6528/ad1df4}
+  doi={10.1088/1361-6528/ad1df4}
 }
 
 @article{Akdeniz2018,
