### 2024-02-09 74c9ea3
Numerous updates and corrections to JOSS manuscript (#30)
* Removed stale links and referenced package website in paper.md

* JOSS Manuscript corrections and updated Tino reference
