+++ publications/2309_JOSS/paper.md
@@ -23,14 +23,14 @@ authors:
 affiliations:
  - name: Department of Chemical Engineering, University of Waterloo, Ontario, Canada
    index: 1
- - name: Department of Physics and Astronomy, University of Waterloo, Ontario, Canada
-   index: 4
- - name: Waterloo Institute for Nanotechnology, University of Waterloo, Ontario, Canada
-   index: 5
  - name: Google Inc.
    index: 2
  - name: East Coast Asset Management SEZC
    index: 3
+ - name: Department of Physics and Astronomy, University of Waterloo, Ontario, Canada
+   index: 4
+ - name: Waterloo Institute for Nanotechnology, University of Waterloo, Ontario, Canada
+   index: 5
 date:
 bibliography: paper.bib
 ---
@@ -43,15 +43,15 @@ The `shapelets` software package provides reference implementations and document
 
 For ease of use, `shapelets` also provides a text configuration-based user interface and Python [entry points](https://packaging.python.org/en/latest/specifications/entry-points/) (custom terminal commands) to improve accessibility for a broad range of potential users in science and engineering, including those without a strong Python programming background. For example, the configuration-file based interface can be invoked via `shapelets config`, and running the unit tests associated with the package can be invoked via `shapelets-test`.
 
-Lastly, the `shapelets` package includes a set of detailed examples which demonstrate usage of the software through both the text configuration-based and programmatic interfaces. These examples include both astronomy and self-assembly applications, providing users with a basis for developing their own applications of the package and shapelets functions, in general.
+Lastly, the `shapelets` package includes a set of detailed examples which demonstrate usage of the software through both the text configuration-based and programmatic interfaces. These examples include both astronomy and self-assembly applications, providing users with a basis for developing their own applications for the package and shapelet functions in general.
 
 # Statement of Need
 
-Shapelets are a class of complete localized orthogonal basis functions with a broad range of applications in image processing and reconstruction [@Refregier2003; @Massey2005; @Suderman2015; @Akdeniz2018; @Tino2024]. Despite their increasingly widespread use, there currently is no single software package that is both broadly accessible (e.g. written in Python or other high-level programming language) and implements several useful applications. Currently, there exists an open-source astronomy-focused [shapelet software package](https://www.astro.dur.ac.uk/~rjm/shapelets/code/index.php) [@Massey2005], however, it is written in the [Interactive Data Language (IDL) programming language](https://en.wikipedia.org/wiki/IDL_(programming_language)) which is not widely used in the science and engineering communities. Furthermore, this package has not been updated in over a decade. Given the increasingly broad usage of shapelets in areas outside of astronomy/astrophysics, an open-source Python-based shapelets software package would provide access to these functions and their applications to a larger community, along with facilitating open-source scientific software development through the existence of a centralized software package that allows for contribution and collaboration.
[…]
