### 2024-03-14 a18c903
Publications (#31)
* Updated JOSS manuscript with accepted shapelet paper info

* Numerous updates and corrections to JOSS manuscript (#30)

* Removed stale links and referenced package website in paper.md

* JOSS Manuscript corrections and updated Tino reference

---------

Co-authored-by: Matthew Tino <mptino@uwaterloo.ca>
