+++ publications/2309_JOSS/paper.md
@@ -23,35 +23,35 @@ authors:
 affiliations:
  - name: Department of Chemical Engineering, University of Waterloo, Ontario, Canada
    index: 1
- - name: Department of Physics and Astronomy, University of Waterloo, Ontario, Canada
-   index: 4
- - name: Waterloo Institute for Nanotechnology, University of Waterloo, Ontario, Canada
-   index: 5
  - name: Google Inc.
    index: 2
  - name: East Coast Asset Management SEZC
    index: 3
+ - name: Department of Physics and Astronomy, University of Waterloo, Ontario, Canada
+   index: 4
+ - name: Waterloo Institute for Nanotechnology, University of Waterloo, Ontario, Canada
+   index: 5
 date:
 bibliography: paper.bib
 ---
 
 # Summary
 
-`Shapelets` is a Python-based software package that implements several shapelet functions [@Refregier2003] and some of their significant applications in science and astronomy. Shapelet functions are a complete and orthogonal set of localized basis functions with mathematical properties convenient for manipulation and analysis of images from a broad range of applications. Over the past few decades, there have been several different shapelet function formulations developed and applied in the areas of astronomy/astrophysics [@Refregier2003; @Massey2005; @Lentati2015; @Birrer2015; @Desvignes2016; @Berge2019], self-assembly nanomaterials [@Suderman2015; @Akdeniz2018; @Tino2023], computational neuroscience [@Victor2006; @Sharpee2009], and medical imaging [@Weissman2004].
+`Shapelets` is a Python-based software package that implements several shapelet functions [@Refregier2003] and some of their significant applications in science and astronomy. Shapelet functions are a complete and orthogonal set of localized basis functions with mathematical properties convenient for manipulation and analysis of images from a broad range of applications. Over the past few decades, there have been several different shapelet function formulations developed and applied in the areas of astronomy/astrophysics [@Refregier2003; @Massey2005; @Lentati2015; @Birrer2015; @Desvignes2016; @Berge2019], self-assembly nanomaterials [@Suderman2015; @Akdeniz2018; @Tino2024], computational neuroscience [@Victor2006; @Sharpee2009], and medical imaging [@Weissman2004].
 
-The `shapelets` software package provides reference implementations and documentation for four different shapelet formulations: cartesian [@Refregier2003], polar [@Massey2005], orthonormal polar with constant radial scale [@Akdeniz2018], and exponential [@Berge2019]. Additionally, the `shapelets` package provides reference implementations of several applications of shapelet functions in astronomy (galactic image decomposition and reconstruction [@Refregier2003; @Massey2005]) and self-assembly (quantification of nanostructure order [@Suderman2015; @Akdeniz2018; @Tino2023]). The coding style of `shapelets` is based on that of [`scipy.special`](https://docs.scipy.org/doc/scipy/reference/special.html) [@scipy].
+The `shapelets` software package provides reference implementations and documentation for four different shapelet formulations: cartesian [@Refregier2003], polar [@Massey2005], orthonormal polar with constant radial scale [@Akdeniz2018], and exponential [@Berge2019]. Additionally, the `shapelets` package provides reference implementations of several applications of shapelet functions in astronomy (galactic image decomposition and reconstruction [@Refregier2003; @Massey2005]) and self-assembly (quantification of nanostructure order [@Suderman2015; @Akdeniz2018; @Tino2024]). The coding style of `shapelets` is based on that of [`scipy.special`](https://docs.scipy.org/doc/scipy/reference/special.html) [@scipy].
 
[…]
