### 2024-03-14 6ce31fb
Update setup.cfg
Updated release version number following JOSS manuscript acceptance
