+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = shapelets
-version = 0.1
+version = 1.0
 description = Shapelets is a module providing both reference access to various shapelet functions and some significant applications of them in science and astronomy.
 long_description = file: README.md, LICENSE
 keywords = shapelets, image analysis
