### 2024-03-14 #6058 @nasserma
Thanks @kyleniemeyer we have:

- Double checked authors and affiliations (including ORCIDs)
- Made a release of the software with the latest changes from the review **(v1.0)**
- Archived the release on Zenodo ([https://doi.org/10.5281/zenodo.10819578](https://doi.org/10.5281/zenodo.10819578))
- Made sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
- Made sure that the license listed for the archive is the same as the software license.

Please let me know if we should do anything else, thanks again!
