### 2024-03-14 #6058 @kyleniemeyer
@nasserma it does look like DOIs might be available for two of those references, can you add?

- Refregier 2003 paper on shapelets: 10.1046/j.1365-8711.2003.05901.x
- Desvignes et al. 2016: High-precision timing of 42 millisecond pulsars with the european pulsar timing array

No need to update the archive, just the paper.
