### 2024-03-14 9291544
Update paper.bib
Added missing DOIs
