+++ publications/2309_JOSS/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={35--47},
   year={2003},
-  publisher={Blackwell Science Ltd Oxford, UK}
+  publisher={Blackwell Science Ltd Oxford, UK},
+  doi={10.1046/j.1365-8711.2003.05901.x}
 }
 
 @article{Massey2005,
@@ -173,14 +174,16 @@
   number={3},
   pages={3341--3380},
   year={2016},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/mnras/stw483}
 }
 
 @book{Wu2012,
   title={Advances in K-means clustering: a data mining thinking},
   author={Wu, Junjie},
   year={2012},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media},
+  doi={10.1007/978-3-642-29807-3}
 }
 
 @incollection{Brock1992,
